+++ frontend/pages/signin/index.tsx
@@ -13,8 +13,8 @@ import styled from "styled-components";
 import { AuthButton } from "../../src/components/atoms/button/AuthButton";
 import { H2Text } from "../../src/components/atoms/text/H2Text";
 import { H3Text } from "../../src/components/atoms/text/H3Text";
-import { InputText } from "../../src/components/atoms/input/Form/InputText";
-import { SubmitButton } from "../../src/components/atoms/input/Form/SubmitButton";
+import { InputText } from "../../src/components/atoms/input/InputText";
+import { SubmitButton } from "../../src/components/atoms/input/SubmitButton";
 import { AuthTemplate } from "../../src/components/template/pages/Auth";
 import { FormWrapper } from "../../src/components/organisms/Form/FormWrapper";
 import { useSignIn } from "../../src/state/ducks/user/operation";
