+++ frontend/src/components/atoms/input/InputFileField.tsx
@@ -1,7 +1,7 @@
 import { memo, ReactNode, VFC } from "react";
 import styled from "styled-components";
-import { color } from "../../../../utility/colors";
-import { device } from "../../../../utility/responsive";
+import { color } from "../../../utility/colors";
+import { device } from "../../../utility/responsive";
 
 type Props = {
   type: string;
