+++ frontend/src/components/atoms/input/InputText.tsx
@@ -1,22 +1,29 @@
 import { memo, ReactNode, VFC } from "react";
-import { useForm } from "react-hook-form";
 import styled from "styled-components";
-import { color } from "../../../../utility/colors";
-import { device } from "../../../../utility/responsive";
+import { color } from "../../../utility/colors";
+import { device } from "../../../utility/responsive";
 
 type Props = {
+  type: string;
+  placeholder?: string;
   children?: ReactNode;
   notice?: string;
-  options: { id: number; name: string }[];
-  // 一旦anyで黙らせる
+  // 一旦anyで黙らせるがstringに戻したい
   register: any;
   errors?: any;
   required?: boolean;
 };
 
-export const InputSelect: VFC<Props> = memo((props) => {
-  const { children, notice, options, errors, required, register } = props;
-  // react-hook-form
+export const InputText: VFC<Props> = memo((props) => {
+  const {
+    type,
+    placeholder,
+    children,
+    notice,
+    register,
+    errors,
+    required,
+  } = props;
   return (
     <>
       <FormItem>
@@ -24,16 +31,16 @@ export const InputSelect: VFC<Props> = memo((props) => {
           {children}
           {required ? <Required>必須</Required> : ""}
         </FormLabel>
-        <Select {...register}>
-          {options &&
-            options.map((option, index) => (
-              <Option id={`${index + 1}`} key={index} value={option.id}>
-                {option.name}
-              </Option>
-            ))}
-        </Select>
+
+        <Input type={type} placeholder={placeholder} {...register} />
         <Notice>{notice}</Notice>
         <ErrorText>
+          {errors[register.name]?.type === "required"
+            ? errors[register.name]?.message
+            : ""}
+          {errors[register.name]?.type === "pattern"
+            ? errors[register.name]?.message
+            : ""}
           {errors[register.name]?.type === "value"
             ? errors[register.name]?.message
             : ""}
@@ -46,10 +53,12 @@ export const InputSelect: VFC<Props> = memo((props) => {
 const FormItem = styled.div`
   text-align: left;
 `;
+
 const FormLabel = styled.div`
   font-weight: bold;
   margin-bottom: 5px;
 `;
+
 const Required = styled.span`
   color: ${color.white};
   background-color: ${color.green};
@@ -66,20 +75,18 @@ const Notice = styled.p`
   margin-top: 5px;
 `;
 
-const Select = styled.select`
+const Input = styled.input`
   width: 100%;
   padding: 14px 10px 14px 16px;
   background-color: ${color.lightGrey};
   color: ${color.black};
   border-radius: 4px;
   text-align: left;
-  cursor: pointer;
   &::placeholder {
     font-size: 16px;
     color: ${color.grey};
   }
 `;
-const Option = styled.option``;
 
 const ErrorText = styled.p`
   color: ${color.red};
